Rights added

+++ controllers/EventController.php
@@ -63,6 +63,11 @@ class EventController extends Controller
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->isGuest){
+            $name = "Permissions";
+            $message = "Vous n'êtes pas authorisé sur cette page";
+            return $this->render('error', ['name' => $name, 'message' => $message]);
+        }
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -75,6 +80,11 @@ class EventController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest){
+            $name = "Permissions";
+            $message = "Vous n'êtes pas authorisé sur cette page";
+            return $this->render('error', ['name' => $name, 'message' => $message]);
+        }
         $model = new Event();
 
         if ($this->request->isPost) {
@@ -112,6 +122,11 @@ class EventController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest){
+            $name = "Permissions";
+            $message = "Vous n'êtes pas authorisé sur cette page";
+            return $this->render('error', ['name' => $name, 'message' => $message]);
+        }
         $model = $this->findModel($id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -132,6 +147,12 @@ class EventController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest){
+            $name = "Permissions";
+            $message = "Vous n'êtes pas authorisé sur cette page";
+            return $this->render('error', ['name' => $name, 'message' => $message]);
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
